backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.36
alembic==1.13.3
asyncpg==0.29.0
redis[hiredis]==5.1.1
minio==7.2.9
hvac==2.3.0
pynacl==1.5.0
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
webauthn==2.0.0
pyhanko==0.25.1
celery[redis]==5.4.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
structlog==24.4.0
argon2-cffi==23.1.0
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
sentry-sdk[fastapi]>=1.39.0
resend>=2.0.0
jinja2>=3.1.0
premailer>=3.10.0
